+++ plugins/markdownexport/apps.py
@@ -1,0 +1,9 @@
+from sysreptor.plugins import PluginConfig
+
+
+class MarkdownExportPluginConfig(PluginConfig):
+    """
+    Export projects as plain markdown documents.
+    """
+    plugin_id = 'ba67f081-bd60-426c-9830-90e71fbce844'
+