+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Fix `run_in_background` after `asgiref` update
 * Log backup started and finished times
 * Restore backup: bulk insert DB objects
+* PDF viewer: add stricter postMessage validation to prevent errors on unexpected messages
 
 
 ## v2025.64 - 2025-07-30